Struct tokio_dns::Connector

pub struct Connector<R> where R: Clone + Resolver {
    // some fields omitted
}

A helper for creating connections.

This object is a wrapper around a LoopHandle and a resolver. It helps initiate connections using endpoints, and offers bultin address translation.

Methods

impl<R> Connector<R> where R: Clone + Resolver

fn new(handle: LoopHandle, resolver: R) -> Self

Create a new Connector.

The handle can be obtained using the handle() method of the Loop object. The resolver_threads parameter is the amount of threads given to the resolver's thread pool. The mode parameter tells the connector how to attempt connection when the resolver yields more than one IP address.